\name{blockX}
\alias{blockX}
\title{Creates a block-diagonal matrix}
\description{
Converts an N by p matrix into a block-diagonal N*K by p*K matrix
}
\usage{
   blockX(X, p, K)
}

\arguments{
   \item{X}{input matrix, of dimensions nobs x nvars)
   \item{K}{number of tasks}
}

\details{
}

\value{
   An (N*K) by (p*K) matrix
}

\references{}

\author{Gad Abraham\cr
Maintainer: Gad Abraham \email{gad.abraham@unimelb.edu.au}
}

\examples{
}

