% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{fmp_api_key}
\alias{fmp_api_key}
\title{sets fmp api key}
\usage{
fmp_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{your api key}
}
\description{
sets fmp api key
}
\examples{
fmp_api_key("demo")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
