% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_statements.R
\name{fmp_balance_sheet}
\alias{fmp_balance_sheet}
\title{Comapny Balance Sheet Items}
\usage{
fmp_balance_sheet(ticker = "AAPL", period = "quarter")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
Comapny Balance Sheet Items
}
\examples{
fmp_balance_sheet(ticker = "AAPL", period = "quarter")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
