% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_statements.R
\name{fmp_cash_flow_statement}
\alias{fmp_cash_flow_statement}
\title{company cash flow statement items}
\usage{
fmp_cash_flow_statement(ticker = "AAPL", period = "quarter")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
company cash flow statement items
}
\examples{
fmp_cash_flow_statement(ticker = "AAPL", period = "quarter")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
