% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_company_profile.R
\name{fmp_company_profile}
\alias{fmp_company_profile}
\title{company profile}
\usage{
fmp_company_profile(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
company profile
}
\examples{
fmp_company_profile(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
