% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fmp_discounted_cash_flow}
\alias{fmp_discounted_cash_flow}
\title{current discount cash flow valuation}
\usage{
fmp_discounted_cash_flow(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
current discount cash flow valuation
}
\examples{
fmp_discounted_cash_flow(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
