% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_enterprise_values.R
\name{fmp_enterprise_values}
\alias{fmp_enterprise_values}
\title{historical enterprise value}
\usage{
fmp_enterprise_values(ticker = "AAPL", period = "quarter")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
historical enterprise value
}
\examples{
fmp_enterprise_values(ticker = "AAPL", period = "quarter")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
