% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etf-holdings.R
\name{fmp_etf_holdings}
\alias{fmp_etf_holdings}
\title{etf holdings information}
\usage{
fmp_etf_holdings(symbol = "IWF")
}
\arguments{
\item{symbol}{etf symbol}
}
\description{
etf holdings information
}
\examples{
fmp_etf_holdings(symbol = "SPY")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
