% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fmp_financial_growth}
\alias{fmp_financial_growth}
\title{historical growth metrics}
\usage{
fmp_financial_growth(ticker = "AAPL", period = "quarter")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
historical growth metrics
}
\examples{
fmp_financial_growth(ticker = "AAPL", period = "quarter")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
