% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fmp_historical_discounted_cash_flow}
\alias{fmp_historical_discounted_cash_flow}
\title{historical discounted cash flow valuation}
\usage{
fmp_historical_discounted_cash_flow(ticker = "AAPL", period = "quarter")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
historical discounted cash flow valuation
}
\examples{
fmp_historical_discounted_cash_flow(ticker = "AAPL", period = "quarter")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
