% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_historical_earning_calendar.R
\name{fmp_historical_earning_calendar}
\alias{fmp_historical_earning_calendar}
\title{historical earnings calendar - premium only}
\usage{
fmp_historical_earning_calendar(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
historical earnings calendar - premium only
}
\examples{
fmp_historical_earning_calendar(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
