% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fmp_historical_employment}
\alias{fmp_historical_employment}
\title{Company historical employment}
\usage{
fmp_historical_employment(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
Company historical employment
}
\examples{
fmp_historical_employment(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
