% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fmp_historical_market_cap}
\alias{fmp_historical_market_cap}
\title{Market Cap}
\usage{
fmp_historical_market_cap(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
Market Cap
}
\examples{
fmp_historical_market_cap(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
