% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fmp_historical_sentiment}
\alias{fmp_historical_sentiment}
\title{Company historical sentiment}
\usage{
fmp_historical_sentiment(ticker = "AAPL", page = 0)
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
Company historical sentiment
}
\examples{
fmp_historical_sentiment(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
