% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_intraday.R
\name{fmp_intraday}
\alias{fmp_intraday}
\title{intraday stock prices}
\usage{
fmp_intraday(
  ticker = "AAPL",
  freq = c("1min", "5min", "15min", "30min", "1hour", "4hour")
)
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{intraday period}
}
\description{
intraday stock prices
}
\examples{
fmp_intraday(ticker = "AAPL", freq = c("1min", "5min", "15min",
    "30min", "1hour", "4hour"))
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
