% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fmp_key_metrics}
\alias{fmp_key_metrics}
\title{company key metrics}
\usage{
fmp_key_metrics(ticker = "AAPL", period = "quarter")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
company key metrics
}
\examples{
fmp_key_metrics(ticker = "AAPL", period = "quarter")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
