% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_prices.R
\name{fmp_quote}
\alias{fmp_quote}
\title{Real Time Stock Details}
\usage{
fmp_quote(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
Real Time Stock Details
}
\examples{
fmp_quote(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
