% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_quote_short.R
\name{fmp_quote_short}
\alias{fmp_quote_short}
\title{Real time stock price}
\usage{
fmp_quote_short(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
Real time stock price
}
\examples{
fmp_quote_short(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
