% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_rating.R
\name{fmp_rating}
\alias{fmp_rating}
\title{Company financial rating}
\usage{
fmp_rating(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
Company financial rating
}
\examples{
fmp_rating(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
