% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fmp_ratios}
\alias{fmp_ratios}
\title{Company financial ratios}
\usage{
fmp_ratios(ticker = "AAPL", period = "quarter")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
Company financial ratios
}
\examples{
fmp_ratios(ticker = "AAPL", period = "quarter")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
