% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fmp_shares_float}
\alias{fmp_shares_float}
\title{Company Share Floats}
\usage{
fmp_shares_float(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}

\item{period}{reporting period, quarter or annual}
}
\description{
Company Share Floats
}
\examples{
fmp_shares_float(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
