% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_prices.R
\name{fmp_stock_dividend}
\alias{fmp_stock_dividend}
\title{company dividend info}
\usage{
fmp_stock_dividend(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
company dividend info
}
\examples{
fmp_stock_dividend(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
