% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_stock_split.R
\name{fmp_stock_split}
\alias{fmp_stock_split}
\title{company stock split information}
\usage{
fmp_stock_split(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
company stock split information
}
\examples{
fmp_stock_split(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
