% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{fmp_analyst_upgrades}
\alias{fmp_analyst_upgrades}
\title{Company financial ratios}
\usage{
fmp_analyst_upgrades(ticker = "AAPL")
}
\arguments{
\item{ticker}{a stock ticker}
}
\description{
Company financial ratios
}
\examples{
fmp_analyst_upgrades(ticker = "AAPL")
}
\seealso{
https://financialmodelingprep.com/developer/docs/
}
